+++ config.py
@@ -25,6 +25,17 @@ class Units(Enum):
     MILLION = 1000 * 1000
 
 
+class FootnoteDataTypes(Enum):
+    EXPENSE = 'expense'
+    INVENTORY = 'inventory'
+
+
+class FootnoteDataSjDivs(Enum):
+    EMPLOYEE_STATUS = 'Employee status'
+    EXPENSE = 'Expense detail'
+    INVENTORY = 'Inventory detail'
+
+
 class AccountDetail(TypedDict):
     names: List[str]
     ids: List[str]
