+++ reports.py
@@ -14,6 +14,7 @@ from config import AccountDetail
 from config import DartResponse
 from config import ReportCodes
 from config import ReportTypes
+from config import FootnoteDataSjDivs
 from config import Units
 from corps import Corp
 from utils import get_api_key
@@ -119,6 +120,7 @@ class Report:
             for is_regular in [True, False]:
                 data.append(
                     {
+                        "sj_div": FootnoteDataSjDivs.EMPLOYEE_STATUS.name,
                         "sj_nm": "직원 등의 현황",
                         "account_nm": f'{team}_{sex}({"정규직" if is_regular else "계약직"})',
                         "amount": (
@@ -163,19 +165,23 @@ class Report:
         return f"{viewer_url}rcpNo={target[2]}&dcmNo={target[3]}&eleId={target[4]}&offset={target[5]}&length={target[6]}&dtd={target[7]}"
 
     def get_footnote_detail_df(
-        self, detail_data_type, unit: Units = Units.DEFAULT
+        self, footnote_data_sj_div: FootnoteDataSjDivs, unit: Units = Units.DEFAULT
     ) -> pd.DataFrame:
-        res = requests.get(self.get_footnote_url())
+        footnote_url = self.get_footnote_url()
+
+        if not footnote_url:
+            return pd.DataFrame()
+        res = requests.get(footnote_url)
         soup = bs(res.text, "html.parser")
 
         target_header = None
-        if detail_data_type == "expense":
+        if footnote_data_sj_div == FootnoteDataSjDivs.EXPENSE:
             pair = re.compile(r"\d+\. 비용")
-            sj_div = "expense_detail"
+            sj_div = FootnoteDataSjDivs.EXPENSE.name
             sj_nm = "비용의 성격별 분류"
-        elif detail_data_type == "inventory":
+        elif footnote_data_sj_div == FootnoteDataSjDivs.INVENTORY:
             pair = re.compile(r"\d+\. 재고자산")
-            sj_div = "inventory_detail"
+            sj_div = FootnoteDataSjDivs.INVENTORY.name
             sj_nm = "재고자산 내역"
         else:
             raise ValueError("Not proper detail_data_type")
@@ -210,7 +216,7 @@ class Report:
             unit_num = 1000 * 1000
         else:
             raise ValueError(
-                f"Invalid unit in inventory detail in {self.corp_name} {self.report_code}"
+                f"Invalid unit in footnote data in {self.corp_name} {self.report_code} {footnote_data_sj_div.name}"
             )
 
         # 단위 조정
@@ -245,10 +251,12 @@ class Report:
             amount = row.find_all("td")[col_index].text.strip().replace(",", "")
 
             try:
-                amount = eval(amount)
+                amount_num = eval(amount)
                 # negative value
                 if re.match(r"\(\d+\)", amount):
-                    amount *= -1
+                    amount = -1 * amount_num
+                else:
+                    amount = amount_num
 
             except SyntaxError:
                 amount = 0
@@ -360,211 +368,3 @@ class Report:
                 }
             )
         return pd.DataFrame(rows)
-
-
-class ReportCalculator:
-    def __init__(
-        self,
-        corp_name: str = None,
-        corp_code: str = None,
-        is_connected: bool = False,
-        unit: Units = Units.DEFAULT,
-        api_key: str = API_KEY,
-    ):
-        if not corp_code and not corp_name:
-            raise ValueError("Either corp_name or corp_code should be vaild")
-
-        corp_inst = Corp(api_key=api_key)
-        if corp_code:
-            target_corp = corp_inst.find_by_code(code=corp_code)
-        if corp_name:
-            target_corp = corp_inst.find_by_name(name=corp_name)
-
-        if not target_corp:
-            raise ValueError("Invalid corp_code")
-
-        self.corp_code = target_corp["corp_code"]
-        self.corp_name = target_corp["corp_name"]
-        self.is_connected = is_connected
-        self.unit = unit
-        self.api_key = None
-        if api_key:
-            self.api_key = api_key
-
-    def refine_unit(self, df):
-        for col in df.columns:
-            if df[col].dtype == int:
-                df[col] = df[col].apply(lambda val: int(val / self.unit.value))
-
-        return df
-
-    def get_annual_data(
-        self, year: int, by_quarter: bool = True, is_accumulated: bool = False
-    ):
-        """
-        :param year
-        :param is_accumulated: True -> 별도의 처리없이 누적값 리턴
-        :param by_quarter: False -> 연간사업보고서 값만 리턴, True -> 분기별 보고서 리턴
-        :return:
-        """
-        # 연간사업보고서 정보만 취합
-        if not by_quarter:
-            report = Report(
-                corp_code=self.corp_code,
-                year=year,
-                report_code=ReportCodes.Q4,
-                is_connected=self.is_connected,
-                api_key=self.api_key,
-            )
-
-            annual_df = pd.DataFrame()
-            for report_type in ReportTypes:
-                target_df = report.get_target_type_data(report_type=report_type)
-                annual_df = pd.concat([annual_df, target_df])
-
-            if annual_df.empty:
-                return pd.DataFrame()
-
-            annual_df.rename(columns={"amount": str(year)}, inplace=True)
-            return self.refine_unit(annual_df)
-
-        # 분기별 정보 취합
-        annual_df = pd.DataFrame()
-
-        # 분기별 컬럼명 저장
-        amount_cols = []
-
-        for i, report_code in enumerate(ReportCodes):
-            amount_col_name = f"{str(year)}.{report_code.name}"
-            report = Report(
-                corp_code=self.corp_code,
-                year=year,
-                report_code=report_code,
-                is_connected=self.is_connected,
-                api_key=self.api_key,
-            )
-
-            quarter_df = pd.DataFrame()
-
-            # 분기별 재무상태표, 손익계산서, 현금흐름표 정보 취합
-            for report_type in ReportTypes:
-                target_df = report.get_target_type_data(report_type=report_type)
-                quarter_df = pd.concat([quarter_df, target_df])
-
-            if quarter_df.empty:
-                continue
-
-            # 분기 데이터가 있을 때에만 컬럼명 저장
-            amount_cols.append(amount_col_name)
-            quarter_df.rename(columns={"amount": amount_col_name}, inplace=True)
-
-            # 1분기 데이터이거나 혹은 이전 분기까지 데이터가 없었던 경우(ex. Q2부터 사업보고서 나오기 시작한 case)
-            if i == 0 or annual_df.empty:
-                annual_df = quarter_df.copy()
-
-            # 분기 데이터를 새 열에 추가
-            else:
-                annual_df[amount_col_name] = quarter_df[amount_col_name]
-
-        # 누적 데이터인 경우 별도의 처리 없이 바로 return
-        if is_accumulated:
-            return self.refine_unit(annual_df)
-
-        if annual_df.empty:
-            return pd.DataFrame()
-
-        # 손익계산서, 현금흐름표는 누적값이므로 별도 처리
-        bs_df = annual_df[annual_df.sj_div == "BS"].copy()
-        leftover_df = annual_df[annual_df.sj_div != "BS"].copy()
-
-        reversed_cols = list(reversed(amount_cols))
-        for i, col in enumerate(reversed_cols):
-            try:
-                prev_quarter_col = reversed_cols[i + 1]
-                leftover_df[col] = leftover_df[col] - leftover_df[prev_quarter_col]
-            except IndexError:
-                pass
-        merged = pd.concat([bs_df, leftover_df])
-        return self.refine_unit(merged)
-
-    def get_annual_data_by_period(
-        self, start_year: int, end_year: int, by_quarter=True, is_accumulated=False
-    ):
-        total_df = pd.DataFrame()
-        join_on_columns = ["sj_div", "sj_nm", "account_nm"]
-
-        for i, year in enumerate(range(start_year, end_year + 1)):
-            annual_data = self.get_annual_data(
-                year=year, by_quarter=by_quarter, is_accumulated=is_accumulated
-            )
-            if i == 0:
-                total_df = annual_data.copy()
-            else:
-                total_df = pd.merge(
-                    total_df,
-                    annual_data,
-                    left_on=join_on_columns,
-                    right_on=join_on_columns,
-                )
-
-        # Drop unused column
-        total_df.drop(["sj_div"], axis=1, inplace=True)
-
-        return total_df
-
-    def write_data(
-        self,
-        start_year: int,
-        end_year: int,
-        is_accumulated: bool = False,
-        filename=None,
-        cell_width=None,
-    ):
-        df_by_quarter = self.get_annual_data_by_period(
-            start_year=start_year,
-            end_year=end_year,
-            is_accumulated=is_accumulated,
-            by_quarter=True,
-        )
-
-        df_by_year = self.get_annual_data_by_period(
-            start_year=start_year,
-            end_year=end_year,
-            is_accumulated=is_accumulated,
-            by_quarter=False,
-        )
-
-        if not filename:
-            filename = f"{self.corp_name}_{str(start_year)}_{str(end_year)}_unit_{self.unit.name.lower()}.xlsx"
-
-        # Formatting cell width in excel file
-        if not cell_width:
-            if self.unit == Units.DEFAULT:
-                cell_width = 15
-            elif self.unit == Units.THOUSAND:
-                cell_width = 12
-            else:
-                cell_width = 9
-
-        with pd.ExcelWriter(filename, engine="xlsxwriter") as writer:
-            df_by_quarter.to_excel(
-                writer,
-                sheet_name="Quarter",
-                index=False,
-                header=True,
-                float_format="#,###",
-            )
-            df_by_year.to_excel(
-                writer,
-                sheet_name="Year",
-                index=False,
-                header=True,
-                float_format="#,###",
-            )
-
-            workbook = writer.book
-            float_format = workbook.add_format({"num_format": "#,##0"})
-            for worksheet in [writer.sheets["Quarter"], writer.sheets["Year"]]:
-                worksheet.set_column(
-                    0, 1000, width=cell_width, cell_format=float_format
-                )
